pkg/proc,pkg/debug: Move core_test to debug package
